refactor: move parse_lexical into subpackage like others
